wslにwindowsのpathを引き継がないようにした

+++ init.sh
@@ -3,6 +3,14 @@ set -xe
 
 # bash -c "$(curl -fsLS https://raw.githubusercontent.com/2lu3/dotfiles/main/init.sh)"
 
+if [ "`echo $PATH | grep '/mnt/c/'`" ]; then
+    # wslでwindowsのpathを引き継がないようにする
+    sudo sh -c "(echo \"[interop]\"; echo \"appendWindowsPath = false\") > /etc/wsl.conf"
+    echo "using windows path from wsl is disabled now"
+    echo "please run 'wsl --shutdown' in your command prompt and run this code again"
+    exit()
+fi
+
 rm ~/.local/share/chezmoi -rf
 mkdir -p ~/.local/
